Refactor CreateAndUpload into ResourcesUpload
