+++ packages/web-app-files/src/helpers/resource/actions/index.ts
@@ -1,2 +1,3 @@
 export * from './transfer'
+export * from './upload'
 export * from './types'
