+++ packages/web-app-files/src/helpers/resource/index.ts
@@ -1,5 +1,5 @@
-export * from './actions'
 export * from './filter'
 export * from './resource'
 export * from './sameResource'
 export * from './conflictHandling'
+export * from './actions'
